+++ src/snowcli/cli/snowpark_shared.py
@@ -7,23 +7,18 @@ from typing import List
 
 import click
 import typer
-from snowcli import utils
-from snowcli.utils import (
-    YesNoAskOptionsType,
-    yes_no_ask_callback,
-)
-from snowcli.zipper import zip_dir
+from requirements.requirement import Requirement
+from snowcli.utils import package_utils
+from snowcli.utils.models import PypiOption
+from snowcli.utils.zipper import zip_dir
 
-PyPiDownloadOption = typer.Option(
-    "ask",
-    help="Whether to download non-Anaconda packages from PyPi. Valid values include: `yes`, `no`, `ask`. Default: `no`.",
-    callback=yes_no_ask_callback,
+PyPiDownloadOption: PypiOption = typer.Option(
+    PypiOption.ASK.value, help="Whether to download non-Anaconda packages from PyPi."
 )
 
-PackageNativeLibrariesOption = typer.Option(
-    "ask",
-    help="When using packages from PyPi, whether to allow native libraries. Valid values include: `yes`, `no`, `ask`. Default: `no`.",
-    callback=yes_no_ask_callback,
+PackageNativeLibrariesOption: PypiOption = typer.Option(
+    PypiOption.NO.value,
+    help="When using packages from PyPi, whether to allow native libraries.",
 )
 
 CheckAnacondaForPyPiDependencies: bool = typer.Option(
@@ -49,63 +44,63 @@ OverwriteOption = typer.Option(
 
 log = logging.getLogger(__name__)
 
+REQUIREMENTS_SNOWFLAKE = "requirements.snowflake.txt"
+REQUIREMENTS_OTHER = "requirements.other.txt"
+
 
 def snowpark_package(
     source: Path,
     artefact_file: Path,
-    pypi_download: YesNoAskOptionsType,
+    pypi_download: PypiOption,
     check_anaconda_for_pypi_deps: bool,
-    package_native_libraries: YesNoAskOptionsType,
+    package_native_libraries: PypiOption,
 ):
     log.info("Resolving any requirements from requirements.txt...")
-    requirements = utils.parse_requirements()
+    requirements = package_utils.parse_requirements()
     if requirements:
         log.info("Comparing provided packages from Snowflake Anaconda...")
-        split_requirements = utils.parse_anaconda_packages(requirements)
+        split_requirements = package_utils.parse_anaconda_packages(requirements)
         if not split_requirements.other:
             log.info("No packages to manually resolve")
-        if split_requirements.other:
-            log.info("Writing requirements.other.txt...")
-            with open("requirements.other.txt", "w", encoding="utf-8") as f:
-                for package in split_requirements.other:
-                    f.write(package.line + "\n")
-        # if requirements.other.txt exists
-        if os.path.isfile("requirements.other.txt"):
+        else:
+            _write_requirements_file(REQUIREMENTS_OTHER, split_requirements.other)
             do_download = (
                 click.confirm(
                     "Do you want to try to download non-Anaconda packages?",
                     default=True,
                 )
-                if pypi_download == "ask"
-                else pypi_download == "yes"
+                if pypi_download == PypiOption.ASK
+                else pypi_download == PypiOption.YES
             )
             if do_download:
                 log.info("Installing non-Anaconda packages...")
-                should_pack, second_chance_results = utils.install_packages(
-                    "requirements.other.txt",
+                should_continue, second_chance_results = package_utils.install_packages(
+                    REQUIREMENTS_OTHER,
                     check_anaconda_for_pypi_deps,
                     package_native_libraries,
                 )
-                if should_pack:
-                    # add the Anaconda packages discovered as dependencies
-                    if second_chance_results is not None:
-                        split_requirements.snowflake = (
-                            split_requirements.snowflake
-                            + second_chance_results.snowflake
-                        )
+                # add the Anaconda packages discovered as dependencies
+                if should_continue and second_chance_results:
+                    split_requirements.snowflake = (
+                        split_requirements.snowflake + second_chance_results.snowflake
+                    )
 
         # write requirements.snowflake.txt file
         if split_requirements.snowflake:
-            log.info("Writing requirements.snowflake.txt file...")
-            with open(
-                "requirements.snowflake.txt",
-                "w",
-                encoding="utf-8",
-            ) as f:
-                for package in utils.deduplicate_and_sort_reqs(
-                    split_requirements.snowflake
-                ):
-                    f.write(package.line + "\n")
+            _write_requirements_file(
+                REQUIREMENTS_SNOWFLAKE,
+                package_utils.deduplicate_and_sort_reqs(split_requirements.snowflake),
+            )
 
     zip_dir(source=source, dest_zip=artefact_file)
-    log.info("Deployment package now ready: app.zip")
+
+    if Path(".packages").exists():
+        zip_dir(source=Path(".packages"), dest_zip=artefact_file, mode="a")
+    log.info(f"Deployment package now ready: %s", artefact_file)
+
+
+def _write_requirements_file(file_name: str, requirements: List[Requirement]):
+    log.info(f"Writing %s file", file_name)
+    with open(file_name, "w", encoding="utf-8") as f:
+        for req in requirements:
+            f.write(f"{req.line}\n")