+++ src/snowcli/utils/zipper.py
@@ -2,9 +2,8 @@ from __future__ import annotations
 
 import fnmatch
 import logging
-import os
 from pathlib import Path
-from typing import Generator, Iterator, List
+from typing import Iterator, Literal
 from zipfile import ZIP_DEFLATED, ZipFile
 
 log = logging.getLogger(__name__)
@@ -36,15 +35,17 @@ def add_file_to_existing_zip(zip_file: str, file: str):
         file (str): The new file to add
     """
     with ZipFile(zip_file, mode="a") as myzip:
-        myzip.write(file, os.path.basename(file))
+        myzip.write(file, Path(file).name)
 
 
-def zip_dir(source: Path, dest_zip: Path) -> None:
+def zip_dir(
+    source: Path, dest_zip: Path, mode: Literal["r", "w", "x", "a"] = "w"
+) -> None:
     files_to_pack: Iterator[Path] = filter(
         _to_be_zipped, map(lambda f: f.absolute(), source.glob("**/*"))
     )
 
-    with ZipFile(dest_zip, "w", ZIP_DEFLATED, allowZip64=True) as package_zip:
+    with ZipFile(dest_zip, mode, ZIP_DEFLATED, allowZip64=True) as package_zip:
         for file in files_to_pack:
             log.debug("Adding %s to %s", file, dest_zip)
             package_zip.write(file, arcname=file.relative_to(source.absolute()))