+++ tests/nativeapp/test_package_scripts.py
@@ -1,35 +1,35 @@
-from pathlib import Path
 from textwrap import dedent
 from unittest import mock
-from unittest.mock import PropertyMock
 
 import pytest
-from snowcli.cli.nativeapp.manager import (
+from snowcli.cli.nativeapp.exceptions import (
     InvalidPackageScriptError,
     MissingPackageScriptError,
-    NativeAppManager,
 )
+from snowcli.cli.nativeapp.run_processor import NativeAppRunProcessor
+from snowcli.cli.project.definition_manager import DefinitionManager
 from snowflake.connector import ProgrammingError
 
+from tests.nativeapp.patch_utils import mock_connection
+from tests.nativeapp.utils import (
+    NATIVEAPP_MANAGER_EXECUTE,
+    NATIVEAPP_MANAGER_EXECUTE_QUERIES,
+)
 from tests.project.fixtures import *
 from tests.testing_utils.fixtures import *
 
-NATIVEAPP_MODULE = "snowcli.cli.nativeapp.manager"
-NATIVEAPP_MANAGER_EXECUTE_QUERIES = (
-    f"{NATIVEAPP_MODULE}.NativeAppManager._execute_queries"
-)
-NATIVEAPP_MANAGER_EXECUTE_QUERY = f"{NATIVEAPP_MODULE}.NativeAppManager._execute_query"
 
-
-mock_connection = mock.patch(
-    "snowcli.cli.common.cli_global_context._CliGlobalContextAccess.connection",
-    new_callable=PropertyMock,
-)
+def _get_na_manager(working_dir):
+    dm = DefinitionManager(working_dir)
+    return NativeAppRunProcessor(
+        project_definition=dm.project_definition["native_app"],
+        project_root=dm.project_root,
+    )
 
 
 @mock.patch(NATIVEAPP_MANAGER_EXECUTE_QUERIES)
-@mock.patch(NATIVEAPP_MANAGER_EXECUTE_QUERY)
-@mock_connection
+@mock.patch(NATIVEAPP_MANAGER_EXECUTE)
+@mock_connection()
 @pytest.mark.parametrize(
     "project_definition_files,expected_call",
     [
@@ -47,7 +47,7 @@ def test_package_scripts(
 ):
     mock_conn.return_value = MockConnectionCtx()
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     native_app_manager._apply_package_scripts()
     assert mock_execute_query.mock_calls == [
         mock.call(expected_call),
@@ -85,7 +85,7 @@ def test_package_scripts(
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_missing_package_script(mock_execute, project_definition_files):
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     with pytest.raises(MissingPackageScriptError):
         (working_dir / "002-shared.sql").unlink()
         native_app_manager._apply_package_scripts()
@@ -98,7 +98,7 @@ def test_missing_package_script(mock_execute, project_definition_files):
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_invalid_package_script(mock_execute, project_definition_files):
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     with pytest.raises(InvalidPackageScriptError):
         second_file = working_dir / "002-shared.sql"
         second_file.unlink()
@@ -113,7 +113,7 @@ def test_invalid_package_script(mock_execute, project_definition_files):
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_undefined_var_package_script(mock_execute, project_definition_files):
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     with pytest.raises(InvalidPackageScriptError):
         second_file = working_dir / "001-shared.sql"
         second_file.unlink()
@@ -124,8 +124,8 @@ def test_undefined_var_package_script(mock_execute, project_definition_files):
 
 
 @mock.patch(NATIVEAPP_MANAGER_EXECUTE_QUERIES)
-@mock.patch(NATIVEAPP_MANAGER_EXECUTE_QUERY)
-@mock_connection
+@mock.patch(NATIVEAPP_MANAGER_EXECUTE)
+@mock_connection()
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_package_scripts_w_missing_warehouse_exception(
     mock_conn,
@@ -141,7 +141,7 @@ def test_package_scripts_w_missing_warehouse_exception(
     )
 
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
 
     with pytest.raises(ProgrammingError) as err:
         native_app_manager._apply_package_scripts()
@@ -149,8 +149,8 @@ def test_package_scripts_w_missing_warehouse_exception(
     assert "Please provide a warehouse for the active session role" in err.value.msg
 
 
-@mock.patch(NATIVEAPP_MANAGER_EXECUTE_QUERY)
-@mock_connection
+@mock.patch(NATIVEAPP_MANAGER_EXECUTE)
+@mock_connection()
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_package_scripts_w_warehouse_access_exception(
     mock_conn,
@@ -163,7 +163,7 @@ def test_package_scripts_w_warehouse_access_exception(
     )
 
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
 
     with pytest.raises(ProgrammingError) as err:
         native_app_manager._apply_package_scripts()