+++ tests/snowpark/test_function.py
@@ -7,15 +7,17 @@ from snowflake.connector import ProgrammingError
 
 
 @mock.patch("snowflake.connector.connect")
-@mock.patch("snowcli.cli.snowpark.commands.ObjectManager.describe")
+@mock.patch("snowcli.cli.snowpark.commands.ObjectManager")
 def test_deploy_function(
-    mock_describe,
+    mock_object_manager,
     mock_connector,
     mock_ctx,
     runner,
     project_directory,
 ):
-    mock_describe.side_effect = ProgrammingError("does not exist or not authorized")
+    mock_object_manager.return_value.describe.side_effect = ProgrammingError(
+        "does not exist or not authorized"
+    )
     ctx = mock_ctx()
     mock_connector.return_value = ctx
     with project_directory("snowpark_functions") as project_dir:
@@ -37,7 +39,7 @@ def test_deploy_function(
             create or replace function func1(a string, b variant)
             returns string
             language python
-            runtime_version=3.8
+            runtime_version=3.10
             imports=('@dev_deployment/my_snowpark_project/app.zip')
             handler='app.func1_handler'
             packages=()
@@ -47,15 +49,21 @@ def test_deploy_function(
 
 
 @mock.patch("snowflake.connector.connect")
-@mock.patch("snowcli.cli.snowpark.commands.ObjectManager.describe")
+@mock.patch("snowcli.cli.snowpark.commands.ObjectManager")
 def test_deploy_function_with_external_access(
-    mock_describe,
+    mock_object_manager,
     mock_connector,
     mock_ctx,
     runner,
     project_directory,
 ):
-    mock_describe.side_effect = ProgrammingError("does not exist or not authorized")
+    mock_object_manager.return_value.show.return_value = [
+        {"name": "external_1", "type": "EXTERNAL_ACCESS"},
+        {"name": "external_2", "type": "EXTERNAL_ACCESS"},
+    ]
+    mock_object_manager.return_value.describe.side_effect = ProgrammingError(
+        "does not exist or not authorized"
+    )
     ctx = mock_ctx()
     mock_connector.return_value = ctx
 
@@ -90,17 +98,21 @@ def test_deploy_function_with_external_access(
 
 
 @mock.patch("snowflake.connector.connect")
-@mock.patch("snowcli.cli.snowpark.commands.ObjectManager.describe")
+@mock.patch("snowcli.cli.snowpark.commands.ObjectManager")
 def test_deploy_function_secrets_without_external_access(
-    mock_describe,
-    mock_connector,
-    mock_ctx,
+    mock_object_manager,
+    mock_conn,
     runner,
+    mock_ctx,
     project_directory,
+    snapshot,
 ):
-    mock_describe.side_effect = ProgrammingError("does not exist or not authorized")
+    mock_object_manager.return_value.show.return_value = [
+        {"name": "external_1", "type": "EXTERNAL_ACCESS"},
+        {"name": "external_2", "type": "EXTERNAL_ACCESS"},
+    ]
     ctx = mock_ctx()
-    mock_connector.return_value = ctx
+    mock_conn.return_value = ctx
 
     with project_directory("snowpark_function_secrets_without_external_access"):
         result = runner.invoke(
@@ -111,9 +123,7 @@ def test_deploy_function_secrets_without_external_access(
         )
 
     assert result.exit_code == 1, result.output
-    assert result.output.__contains__(
-        "Can not provide secrets without external access integration"
-    )
+    assert result.output == snapshot
 
 
 @mock.patch("snowflake.connector.connect")
@@ -151,7 +161,6 @@ def test_deploy_function_no_changes(
     assert queries == [
         "create stage if not exists dev_deployment comment='deployments managed by snowcli'",
         f"put file://{Path(project_dir).resolve()}/app.zip @dev_deployment/my_snowpark_project auto_compress=false parallel=4 overwrite=True",
-        "describe function func1(string, variant)",
     ]
 
 
@@ -190,13 +199,12 @@ def test_deploy_function_needs_update_because_packages_changes(
     assert queries == [
         "create stage if not exists dev_deployment comment='deployments managed by snowcli'",
         f"put file://{Path(project_dir).resolve()}/app.zip @dev_deployment/my_snowpark_project auto_compress=false parallel=4 overwrite=True",
-        "describe function func1(string, variant)",
         dedent(
             """\
             create or replace function func1(a string, b variant)
             returns string
             language python
-            runtime_version=3.8
+            runtime_version=3.10
             imports=('@dev_deployment/my_snowpark_project/app.zip')
             handler='app.func1_handler'
             packages=('foo=1.2.3','bar>=3.0.0')
@@ -241,13 +249,12 @@ def test_deploy_function_needs_update_because_handler_changes(
         "create stage if not exists dev_deployment comment='deployments managed by snowcli'",
         f"put file://{Path(project_dir).resolve()}/app.zip @dev_deployment/my_snowpark_project"
         f" auto_compress=false parallel=4 overwrite=True",
-        "describe function func1(string, variant)",
         dedent(
             """\
             create or replace function func1(a string, b variant)
             returns string
             language python
-            runtime_version=3.8
+            runtime_version=3.10
             imports=('@dev_deployment/my_snowpark_project/app.zip')
             handler='app.func1_handler'
             packages=('foo=1.2.3','bar>=3.0.0')
@@ -284,18 +291,22 @@ def _deploy_function(
 ):
     ctx = mock_ctx(mock_cursor(rows=rows, columns=[]))
     mock_connector.return_value = ctx
-    with project_directory("snowpark_functions") as temp_dir:
-        (Path(temp_dir) / "requirements.snowflake.txt").write_text(
-            "foo=1.2.3\nbar>=3.0.0"
-        )
-        result = runner.invoke(
-            [
-                "snowpark",
-                "deploy",
-                "--format",
-                "json",
-                *args,
-            ]
-        )
+    with mock.patch("snowcli.cli.snowpark.commands.ObjectManager") as om:
+
+        om.return_value.describe.return_value = rows
+
+        with project_directory("snowpark_functions") as temp_dir:
+            (Path(temp_dir) / "requirements.snowflake.txt").write_text(
+                "foo=1.2.3\nbar>=3.0.0"
+            )
+            result = runner.invoke(
+                [
+                    "snowpark",
+                    "deploy",
+                    "--format",
+                    "json",
+                    *args,
+                ]
+            )
     queries = ctx.get_queries()
     return queries, result, temp_dir