+++ tests_integration/test_snowpark.py
@@ -1,12 +1,12 @@
 from __future__ import annotations
 
+import os.path
 from pathlib import Path
 
 import pytest
 
 from tests_integration.testing_utils import (
     SnowparkTestSteps,
-    assert_that_result_is_successful,
 )
 from tests_integration.testing_utils.snowpark_utils import (
     SnowparkTestSetup,
@@ -90,7 +90,7 @@ def test_snowpark_flow(_test_steps, project_directory, alter_snowflake_yml):
 
         # Subsequent deploy of same object should fail
         _test_steps.snowpark_deploy_should_return_error_with_message_contains(
-            "already exists"
+            "Following objects already exists"
         )
 
         # Apply changes to project objects
@@ -180,6 +180,35 @@ def test_snowpark_flow(_test_steps, project_directory, alter_snowflake_yml):
 
         _test_steps.assert_that_only_these_files_are_staged_in_test_db(
             *expected_files, stage_name=STAGE_NAME
+        )
+
+
+@pytest.mark.integration
+def test_snowpark_with_separately_created_package(
+    _test_steps, project_directory, alter_snowflake_yml
+):
+
+    _test_steps.package_should_build_proper_artifact("PyRTF3")
+    _test_steps.package_should_upload_artifact_to_stage("PyRTF3.zip", STAGE_NAME)
+
+    _test_steps.artifacts_left_after_package_creation_should_be_deleted("PyRTF3.zip")
+
+    with project_directory("snowpark_with_package") as proj_dir:
+        _test_steps.snowpark_build_should_zip_files()
+
+        _test_steps.snowpark_deploy_should_finish_successfully_and_return(
+            [
+                {
+                    "object": "test_func(name string)",
+                    "status": "created",
+                    "type": "function",
+                },
+            ]
+        )
+        _test_steps.snowpark_execute_should_return_expected_value(
+            object_type="function",
+            identifier="test_func('foo')",
+            expected_value="<slot wrapper '__str__' of 'object' objects> foo",
         )